Fix handling nested defines and refactor
